+++ R/protectLinkedTables.R
@@ -1,36 +1,43 @@
-#' protect two \code{\link{sdcProblem-class}} objects that have common cells
+#' Protect two tables with common cells
 #'
-#' \code{\link{protectLinkedTables}} can be used to protect tables, that have
+#' [protectLinkedTables()] can be used to protect tables that have
 #' common cells. It is of course required that after the anonymization process
 #' has finished, all common cells have the same anonymization state in both
 #' tables.
 #'
-#' @param objectA a \code{\link{sdcProblem-class}} object
-#' @param objectB a \code{\link{sdcProblem-class}} object
-#' @param commonCells a list object defining common cells in \code{objectA} and \code{objectB}. For each variable that has one or more common codes in both tables, a list element needs to be specified.
-#' \itemize{
-#' \item List-elements of length 3: Variable has exact same levels and structure in both tables
-#' \itemize{
-#' \item \code{first element}: character vector of length 1 specifying the variable name in argument \code{objectA}
-#' \item \code{second element}: character vector of length 1 specifying the variable name in argument \code{objectB}
-#' \item \code{third element}: character vector of length 1 being with keyword \code{ALL} }
-#' \item List-elements of length 4: Variable has different codes and levels in tables \code{objectA} and \code{objectB}
-#' \itemize{
-#' \item \code{first element}: character vector of length 1 specifying the variable name in argument \code{objectA}
-#' \item \code{second element}: character vector of length 1 specifying the variable name in argument \code{objectB}
-#' \item \code{third element}: character vector defining codes within \code{objectA}
-#' \item \code{fourth element}: character vector with length that equals the length of the third list-element. The vector defines codes of the variable in \code{objectB} that match the codes given in the third list-element for \code{objectA}.
-#' }
-#' }
-#' @param method a character vector of length 1 specifying the algorithm that should be used to protect the primary sensitive table cells. Allowed values are:
-#' \itemize{
-#' \item \code{HITAS}:
-#' \item \code{SIMPLEHEURISTIC}:
-#' \item \code{OPT}: }
-#' @param ... additional arguments to control the secondary cell suppression algorithm. For details, see \code{\link{protectTable}}.
+#' @param objectA a [sdcProblem-class] object
+#' @param objectB a [sdcProblem-class] object
+#' @param commonCells a list object defining common cells in 
+#' `objectA` and `objectB`. For each variable that has one or more common 
+#' codes in both tables, a list element needs to be specified.
+#' - List-elements of length 3: Variable has exact same levels and structure 
+#' in both input tables
+#'    * `first element`: scalar character vector specifying the variable 
+#'    name in argument `objectA`
+#'    * `second element`: scalar character vector specifying the variable 
+#'    name in argument `objectB`
+#'    * `third element`: scalar character vector being with keyword `"ALL"`
+#' - List-elements of length `4`: Variable has different codes and levels 
+#' in tables `objectA` and `objectB`
+#'    * `first element`: scalar character vector specifying the variable 
+#'    name in argument `objectA`
+#'    * `second element`: scalar character vector specifying the variable 
+#'    name in argument `objectB`
+#'    * `third element`: character vector defining codes within `objectA`
+#'    * `fourth element`: character vector with length that equals the length 
+#'    of the third list-element. This vector defines codes of the dimensional 
+#'    variable in `objectB` that match the codes given in the third list-element 
+#'    for `objectA`.
+#' @param method scalar character vector defining the algorithm 
+#' that should be used to protect the primary sensitive table cells. The possible
+#' values are `"HITAS"`, `"SIMPLEHEURISTIC"` and `"OPT"`; For details please see
+#' [protectTable()].
+#' @param ... additional arguments to control the secondary cell suppression 
+#' algorithm. For details, see [protectTable()].
 #'
-#' @return a list of length 2 with each list-element being an \code{\link{safeObj-class}} object
-#'
+#' @return a list of length `2` with each list-element being an 
+#' [safeObj-class] object
+#' @md
 #' @examples
 #' \dontrun{
 #' # load micro data for further processing
@@ -65,20 +72,19 @@
 #' dl1 <- list(gender = d_gender, ecoOld = d_eco1)
 #' dl2 <- list(region = d_region, gender = d_gender, ecoNew = d_eco2)
 #'
-#' # creating input objects for further processing. For details have a look at
-#' # \code{\link{makeProblem}}.
+#' # creating input objects for further processing. 
+#' # For details, see ?makeProblem.
 #' p1 <- makeProblem(
 #'   data = microData1, 
 #'   dimList = dl1, 
 #'   dimVarInd = 1:2,
-#'   numVarInd = 3
-#' )
+#'   numVarInd = 3)
+#'   
 #' p2 <- makeProblem(
 #'   data = microData2, 
 #'   dimList = dl2, 
 #'   dimVarInd = 1:3,
-#'   numVarInd = 4
-#' )
+#'   numVarInd = 4)
 #'
 #' # the cell specified by gender == "Tot" and ecoOld == "A"
 #' # is one of the common cells! -> we mark it as primary suppression
@@ -87,8 +93,7 @@
 #'   characteristics = c("Tot", "A"),
 #'   varNames = c("gender", "ecoOld"), 
 #'   rule = "u", 
-#'   verbose = FALSE
-#' )
+#'   verbose = FALSE)
 #'
 #' # the cell specified by region == "Tot" and gender == "f" and ecoNew == "C"
 #' # is one of the common cells! -> we mark it as primary suppression
@@ -97,8 +102,7 @@
 #'   characteristics = c("Tot", "f", "C"),
 #'   varNames = c("region", "gender", "ecoNew"), 
 #'   rule = "u", 
-#'   verbose = FALSE
-#' )
+#'   verbose = FALSE)
 #'
 #' # specifying input to define common cells
 #' common_cells <- list()
@@ -107,17 +111,20 @@
 #' common_cells$v.gender <- list()
 #' common_cells$v.gender[[1]] <- "gender" # variable name in "p1"
 #' common_cells$v.gender[[2]] <- "gender" # variable name in "p2"
+#' 
 #' # "gender" has equal characteristics on both datasets -> keyword "ALL"
 #' common_cells$v.gender[[3]] <- "ALL"
 #'
 #' # variables: "ecoOld" and "ecoNew"
 #' common_cells$v.eco <- list()
-#' common_cells$v.eco[[1]] <- "ecoOld"   # variable name in "p1"
-#' common_cells$v.eco[[2]] <- "ecoNew"   # variable name in "p2"
+#' common_cells$v.eco[[1]] <- "ecoOld" # variable name in "p1"
+#' common_cells$v.eco[[2]] <- "ecoNew" # variable name in "p2"
 #'
-#' # vector of common characteristics: "A" and "B" in variable "ecoOld" in 'p1'
+#' # vector of common characteristics: 
+#' # "A" and "B" in variable "ecoOld" in "p1"
 #' common_cells$v.eco[[3]] <- c("A", "B")
-#' # correspond to characteristics "C" and "D" in variable "ecoNew" in "p2"
+#' 
+#' # correspond to codes "C" and "D" in variable "ecoNew" in "p2"
 #' common_cells$v.eco[[4]] <- c("C", "D")
 #'
 #' # protect the linked data
@@ -126,22 +133,21 @@
 #'   objectB = p2, 
 #'   commonCells = common_cells, 
 #'   method = "HITAS", 
-#'   verbose = TRUE
-#' )
+#'   verbose = TRUE)
 #'
 #' # having a look at the results
-#' result.tab1 <- result[[1]]
-#' result.tab2 <- result[[2]]
-#' summary(result.tab1)
-#' summary(result.tab2)
+#' result_tab1 <- result[[1]]
+#' result_tab2 <- result[[2]]
+#' summary(result_tab1)
+#' summary(result_tab2)
 #' }
 #'
 #' @rdname protectLinkedTables
 #' @export protectLinkedTables
-#' @seealso \code{\link{protectTable}}
+#' @seealso [protectTable()]
 #' @author Bernhard Meindl \email{bernhard.meindl@@statistik.gv.at}
 protectLinkedTables <- function(objectA, objectB, commonCells, method, ...) {
-  f.calcCommonCellIndices <- function(input1, input2, commonCells) {
+  .indices_common_cells <- function(input1, input2, commonCells) {
     id1 <- id2 <- NULL
     dat1 <- g_df(input1); dat1$id1 <- 1:nrow(dat1)
     dat2 <- g_df(input2); dat2$id2 <- 1:nrow(dat2)
@@ -201,7 +207,7 @@ protectLinkedTables <- function(objectA, objectB, commonCells, method, ...) {
     return(list(commonInd1 = mm$id1, commonInd2 = mm$id2))
   }
 
-  f.checkCommonCells <- function(suppPattern1, suppPattern2, commonCellIndices) {
+  .check_common_cells <- function(suppPattern1, suppPattern2, commonCellIndices) {
     indOK <- TRUE
     if (any(suppPattern1[commonCellIndices[[1]]] != suppPattern2[commonCellIndices[[2]]]))
       indOK <- FALSE
@@ -210,7 +216,7 @@ protectLinkedTables <- function(objectA, objectB, commonCells, method, ...) {
 
   ### arguments ###
   if (!method %in% c("HITAS", "OPT", "SIMPLEHEURISTIC")) {
-    stop("valid methods are 'HITAS', 'OPT' or 'SIMPLEHEURISTIC'!\n")
+    stop("valid methods are 'HITAS', 'OPT' or 'SIMPLEHEURISTIC'!", call. = FALSE)
   }
 
   paraList <- genParaObj(selection = "control.secondary", method = method, ...)
@@ -237,15 +243,15 @@ protectLinkedTables <- function(objectA, objectB, commonCells, method, ...) {
 
   pI.A <- g_problemInstance(outA)
   pI.B <- g_problemInstance(outB)
+  
   # calc original primary suppressions
-
   origPrimSupp1Index <- g_primSupps(pI.A)
   origPrimSupp2Index <- g_primSupps(pI.B)
 
   # no primary suppressions
   if (length(origPrimSupp1Index) + length(origPrimSupp2Index) == 0) {
     if (paraList$verbose) {
-      cat("\n===> no primary suppressions. All common cells have the same anonymity-status! [Finished]\n")
+      message("===> no primary suppressions. all common cells have the same anonymity-status! [finished]")
     }
     outA <- c_finalize(object = outA, input = paraList)
     outB <- c_finalize(object = outB, input = paraList)
@@ -253,17 +259,17 @@ protectLinkedTables <- function(objectA, objectB, commonCells, method, ...) {
   }
 
   # calculate commonCells:
-  commonCellIndices <- f.calcCommonCellIndices(outA, outB, commonCells)
+  commonCellIndices <- .indices_common_cells(outA, outB, commonCells)
 
   # suppression patterns after the first run
   suppPatternA <- g_suppPattern(pI.A)
   suppPatternB <- g_suppPattern(pI.B)
 
-  indOK <- f.checkCommonCells(suppPatternA, suppPatternB, commonCellIndices)
+  indOK <- .check_common_cells(suppPatternA, suppPatternB, commonCellIndices)
   counter <- 1
   if (!indOK) {
     if (paraList$verbose) {
-      cat("we have to start the iterative procedure!\n")
+      message("we have to start the iterative procedure!")
     }
     runInd <- TRUE
     while (runInd) {
@@ -280,24 +286,39 @@ protectLinkedTables <- function(objectA, objectB, commonCells, method, ...) {
             pI.A <- g_problemInstance(outA)
             s_sdcStatus(pI.A) <- list(
               index = index[[j]],
-              vals = rep("u", length(index[[j]]))
-            )
+              vals = rep("u", length(index[[j]])))
+            
+            # we need to set "z" to "s" for the time being
+            zs <- which(g_sdcStatus(pI.A) == "z")
+            if (length(zs) > 0) {
+              s_sdcStatus(pI.A) <- list(
+                index = zs,
+                vals = rep("s", length(zs)))
+            }
             s_problemInstance(outA) <- pI.A
             s_indicesDealtWith(outA) <- NULL
             s_startJ(outA) <- 1
             s_startI(outA) <- 1
-            outA <- c_anon_worker(outA, input = paraList)
+            outA <- c_quick_suppression(outA, input = paraList)$object
           } else {
             pI.B <- g_problemInstance(outB)
             s_sdcStatus(pI.B) <- list(
               index = index[[j]],
-              vals = rep("u", length(index[[j]]))
-            )
+              vals = rep("u", length(index[[j]])))
+
+            # we need to set "z" to "s" for the time being
+            zs <- which(g_sdcStatus(pI.B) == "z")
+            if (length(zs) > 0) {
+              s_sdcStatus(pI.B) <- list(
+                index = zs,
+                vals = rep("s", length(zs)))
+            }                        
+            
             s_problemInstance(outB) <- pI.B
             s_indicesDealtWith(outB) <- NULL
             s_startJ(outB) <- 1
             s_startI(outB) <- 1
-            outB <- c_anon_worker(outB, input = paraList)
+            outB <- c_quick_suppression(outB, input = paraList)$object
           }
         }
       }
@@ -306,7 +327,7 @@ protectLinkedTables <- function(objectA, objectB, commonCells, method, ...) {
       suppPatternB <- g_suppPattern(g_problemInstance(outB))
 
       cbind(suppPatternA[commonCellIndices[[1]]], suppPatternB[commonCellIndices[[2]]])
-      indOK <- f.checkCommonCells(suppPatternA, suppPatternB, commonCellIndices)
+      indOK <- .check_common_cells(suppPatternA, suppPatternB, commonCellIndices)
       if (indOK)
         runInd <- FALSE
       if (counter > paraList$maxIter) {
@@ -318,7 +339,9 @@ protectLinkedTables <- function(objectA, objectB, commonCells, method, ...) {
     }
   }
   if (paraList$verbose) {
-    cat("\n===> all common cells have the same anonymity-state in both tables after", counter, "iterations! [Finished]\n")
+    message(
+      "===> all common cells have the same anonymity-state in both tables after ", 
+      counter, " iterations! [finished]")
   }
   outA <- c_finalize(object = outA, input = paraList)
   outB <- c_finalize(object = outB, input = paraList)