+++ NEWS.md
@@ -1,10 +1,17 @@
+# sdcTable 0.31
+- remove debugging outut
+- fixing typos in vignettes
+- bugfix in `protectLinkedTables()`
+- remove dependency on package `lpSolveAPI`
+- update singleton-detecton procedure by allowing to input a threshold value that must be respected for all simple table rows
+- [todo] check dominance rules (unweighted variables are used)
+
 # sdcTable 0.30
 - various fixes and improvements in `createJJFormat()`
 - ignore sampling weights in case the input in `makeProblem()` is a complete table
 and not microdata
 - `sdcProb2df():` if numeric variables are shown; display their weighted values (in case)
 sampling weights have been specified in `makeProblem`
-- [todo]: check dominance rules (unweighted variables are used)
 
 # sdcTable 0.29
 - change in `primarySuppression()`: for dominance rule it is now only possible to specify underlying variables by name (using argument `numVarName`) and no longer by index to avoid errors